[分離]DMOptionを独立したファイルに
